+++ lib/rubocop.rb
@@ -127,6 +127,7 @@ require 'rubocop/cop/metrics/method_length'
 require 'rubocop/cop/metrics/parameter_lists'
 require 'rubocop/cop/metrics/perceived_complexity'
 
+require 'rubocop/cop/performance/casecmp'
 require 'rubocop/cop/performance/case_when_splat'
 require 'rubocop/cop/performance/count'
 require 'rubocop/cop/performance/detect'